requirements.txt
discord.py==1.3.3
python-dotenv==0.10.3